contrib/rpm: add "00-server-dhcp-client-id.conf"
While this is packaged also in "NetworkManager-config-server.rpm"
sub-package, it's not in "00-server.conf" file. The reason
is that a convenient way to disable configuration from
"/usr/lib/NetworkManager/conf.d", is by putting a (possibly empty)
file into /etc directory with the same name. If the sub-package
only provides one "00-server.conf" file, this is no longer possible
at a granular level.